requirements.txt
# Core LLM and RAG frameworks
langchain==0.1.0
llama-index==0.9.30
transformers==4.36.0
torch>=2.0.0

# Vector databases
faiss-cpu==1.7.4
chromadb==0.4.18
sentence-transformers==2.2.2

# Web frameworks
streamlit==1.29.0
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Data handling
pandas==2.1.4
numpy==1.24.3
pydantic==2.5.0

# Security and governance
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bleach==6.1.0

# Model acceleration (optional)
accelerate==0.25.0
bitsandbytes==0.41.3

# Additional utilities
pydantic-settings==2.1.0
aiofiles==23.2.1
python-json-logger==2.0.7
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
